backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==22.1
django-safedelete==1.4.1
django-encrypted-model-fields==0.7.2
django-auditlog==2.3.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.19
django-countries==7.5
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
cryptography==41.0.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==0.28.1
google-generativeai==0.3.2
anthropic==0.7.4
pyotp==2.8.0
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
